project/requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.3
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
